requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
PyYAML==6.0.3
structlog==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
